pyproject.toml
[project]
name = "fastapi_oracle_project"
version = "0.1.0"
description = "FastAPI + Oracle backend"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "oracledb>=1.4.0",
    "python-dotenv>=1.0.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic[email]>=2.11.7",
    "pyjwt[crypto]>=2.10.1",
    "sqlmodel>=0.0.24",
    "pydantic-settings>=2.10.1",
    "python-multipart>=0.0.20",
    "psycopg>=3.2.9",
    "bcrypt>=4.3.0",
]

[tool.uv]
dev-dependencies = ["pytest>=7.4.0"]
